Add files via upload
A surpasser of an element of an array is a greater element to its right, therefore x[j] is a surpasser of x[i] if i < j and x[i] < x[j]. The surpasser count of an element is the number of surpassers.
Given an array of distinct integers, for each element of the array find its surpasser count i.e. count the number of elements to the right that are greater than that element. - https://practice.geeksforgeeks.org/problems/surpasser-count/0